+++ packages/widgets/src/components/UserProfile/UserProfile.stories.tsx
@@ -1,0 +1,20 @@
+import { Meta, StoryObj } from '@storybook/react';
+import { Avatar } from '@ssa-ui-kit/core';
+import { UserProfile } from './UserProfile';
+
+export default {
+  title: 'Widgets/UserProfile',
+  component: UserProfile,
+} as Meta<typeof UserProfile>;
+
+export const Default: StoryObj<typeof UserProfile> = () => {
+  return (
+    <UserProfile
+      name="Josh Li"
+      email="Josh@gmail.com"
+      trigger={<Avatar size={42} image="https://via.placeholder.com/42x42" />}
+    />
+  );
+};
+
+Default.args = {};
