+++ packages/widgets/src/components/UserProfile/types.ts
@@ -1,6 +1,5 @@
 export interface UserProfileProps {
   name: string;
   email: string;
-  link: string;
   trigger: string | JSX.Element;
 }
